+++ backend/package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "build": "tsc && tsc-alias",
-    "start": "NODE_ENV=production node ./dist/shared/infra/http/server.js",
-    "dev": "NODE_ENV=development ts-node-dev --inspect --transpile-only -r tsconfig-paths/register src/shared/infra/http/server.ts",
+    "start": "NODE_ENV=production node ./dist/index.js",
+    "dev": "NODE_ENV=development ts-node-dev --inspect --transpile-only -r tsconfig-paths/register src/index.ts",
     "typeorm": "ts-node-dev -r tsconfig-paths/register node_modules/.bin/typeorm -d ./src/shared/infra/typeorm/bootstrap.ts",
     "migration:show": "ts-node-dev -r tsconfig-paths/register node_modules/.bin/typeorm migration:show -d ./src/shared/infra/typeorm/bootstrap.ts",
     "migration:run": "ts-node-dev -r tsconfig-paths/register node_modules/.bin/typeorm migration:run -d ./src/shared/infra/typeorm/bootstrap.ts",
