+++ backend/src/modules/Expense/infra/typeorm/repositories/ExpensesMonthRepository.ts
@@ -2,7 +2,7 @@ import { Repository } from 'typeorm';
 
 import { DataSourceConfiguration } from '@shared/infra/typeorm/bootstrap';
 import { IExpensesMonthRepository } from '@modules/Expense/domain/repositories/IExpensesInMonthRepository';
-import { ExpenseMonthMapper } from '../entities/ExpenseInMonth';
+import { ExpenseMonthMapper } from '../entities/ExpenseMonthMapper';
 import { ExpenseMonth } from '@modules/Expense/domain/models/ExpenseMonth';
 
 export class ExpensesMonthRepository implements IExpensesMonthRepository {
