+++ backend/src/shared/infra/typeorm/bootstrap.ts
@@ -5,15 +5,14 @@ import { DataSource } from 'typeorm';
 const isProduction = process.env.NODE_ENV === 'production';
 
 export const DataSourceConfiguration = new DataSource({
-  type: process.env.DB_TYPE ?? 'postgres',
+  type: (process.env.DB_TYPE as 'postgres' | 'mysql') ?? 'postgres',
   host: process.env.DB_HOST ?? 'localhost',
   port: +(process.env.DB_PORT ?? 5432),
   username: process.env.DB_USER ?? 'root',
   password: process.env.DB_PASSWORD ?? 'root',
   database: process.env.DB_DATABASE ?? 'finance-app',
-  synchronize: true,
   logging: true,
-
+  synchronize: false,
   entities: [
     isProduction
       ? 'dist/modules/**/infra/typeorm/entities/*{.ts,.js}'
@@ -21,7 +20,7 @@ export const DataSourceConfiguration = new DataSource({
   ],
   migrations: [
     isProduction
-      ? '../dist/infra/typeorm/migrations/*{.ts,.js}'
-      : '../src/infra/typeorm/migrations/*{.ts,.js}',
+      ? '../dist/shared/infra/typeorm/migrations/*{.ts,.js}'
+      : '../src/shared/infra/typeorm/migrations/*{.ts,.js}',
   ],
 });
